event_listener

Module notify

The Notification trait for specifying notification.

Modules

__private 🔒

Macros

impl_for_numeric_types 🔒

Structs

GenericNotify 🔒

A generic notification.

Traits

IntoNotification

A value that can be converted into a Notification.

Notification

A notification that can be used to notify an Event.

TagProducer 🔒

The producer for a generic notification.

Functions

full_fence 🔒

Equivalent to atomic::fence(Ordering::SeqCst), but in some cases faster.