event_listener

Struct Inner

pub(crate) struct Inner<T> {
    pub(crate) notified: AtomicUsize,
    pub(crate) list: List<T>,
}

Inner state of Event.

Fields
notified: AtomicUsize

The number of notified entries, or usize::MAX if all of them have been notified.

If there are no entries, this value is set to usize::MAX.

list: List<T>

Inner queue of event listeners.

On std platforms, this is an intrusive linked list. On no_std platforms, this is a more traditional Vec of listeners, with an atomic queue used as a backup for high contention.

Implementations

impl<T> Inner<T>

fn with_inner<R>(&self, f: impl FnOnce(&mut Inner<T>) -> R) -> R

pub(crate) fn insert(&self, listener: Pin<&mut Option<Listener<T>>>)

Add a new listener to the list.

pub(crate) fn remove( &self, listener: Pin<&mut Option<Listener<T>>>, propagate: bool, ) -> Option<State<T>>

Remove a listener from the list.

pub(crate) fn notify(&self, notify: impl Notification<Tag = T>) -> usize

Notifies a number of entries.

pub(crate) fn register( &self, listener: Pin<&mut Option<Listener<T>>>, task: TaskRef<'_>, ) -> RegisterResult<T>

Register a task to be notified when the event is triggered.

Returns true if the listener was already notified, and false otherwise. If the listener isn’t inserted, returns None.

impl<T> Inner<T>

pub(crate) fn new() -> Self